Debug

+++ lib/pixels/src/pixels.h
@@ -63,7 +63,7 @@ class PIXELS
     // marshal returns a pointer to an array of pixels and accepts a pointer to a uint8_t array payload with the length of the array, as
     // well as a pointer to an unsigned integer which will be changed to the number of LEDs decoded from the payload. If invalid a NULL will
     // be returned and the value at pixCnt will be set to zero.
-    pixel *marshal(uint8_t *pyld, unsigned len, unsigned *pixCnt);
+    pixel *marshal(uint8_t *pyld, unsigned len, uint16_t *pixCnt);
 
     uint8_t syncWord = 0x0;
